refactor(step-9): extract WebSocketRouter from monolith
WebSocketRouter class in voxd/router.py (774 lines) with all 11
message handlers, auth check, parse helpers, and connection handler.
Handlers take (self, msg, ws) instead of (msg, ws, ctx) — dependencies
accessed through self._synthesis, self._playback, self._music, etc.

DaemonContext stripped of auth_token, client_count, chime_dedup,
once_dedup, and all music delegation properties.

_monolith.py: 1113→506 lines. method_ratio +0.152, max_complexity
13→5, avg_complexity -0.650, module_size -607.
Test count: 1472→1478 (+6). test_voxd.py: 1115→56 lines.
