+++ .oo-audit.jsonl
@@ -26,3 +26,5 @@
 {"ts":"2026-05-14T02:06:41Z","commit":"4a552e5","files_scored":43,"files_improved":1,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.089,0.407],"encapsulation_ratio":[0.056,0.222],"avg_params":[2.43,1.71],"max_complexity":[20.0,13.0],"avg_complexity":[3.71,2.02],"module_size":[1645.0,1121.0],"classes_per_module":[2.0,1.0],"public_attr_violations":[17.0,7.0]}}}
 {"ts":"2026-05-14T04:39:50Z","commit":"9491c2d","files_scored":44,"files_improved":3,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[134.0,136.0]},"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.407,0.435],"encapsulation_ratio":[0.222,0.375],"avg_params":[1.71,1.67],"avg_complexity":[2.03,1.98],"module_size":[1130,1113.0],"public_attr_violations":[7.0,5.0]},"src/punt_vox/voxd/health.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.57],"max_complexity":[0.0,2.0],"avg_complexity":[0.0,1.14],"module_size":[0.0,93.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
 {"ts":"2026-05-14T04:42:03Z","commit":"9491c2d","files_scored":44,"files_improved":3,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[134.0,136.0]},"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.407,0.435],"encapsulation_ratio":[0.222,0.375],"avg_params":[1.71,1.67],"avg_complexity":[2.03,1.98],"module_size":[1130,1113.0],"public_attr_violations":[7.0,5.0]},"src/punt_vox/voxd/health.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,0.57],"max_complexity":[0.0,2.0],"avg_complexity":[0.0,1.14],"module_size":[0.0,93.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T04:56:35Z","commit":"fac9111","files_scored":45,"files_improved":3,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[136.0,122.0]},"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.435,0.587],"avg_params":[1.67,1.43],"max_complexity":[13.0,5.0],"avg_complexity":[1.98,1.33],"module_size":[1113.0,506.0],"class_to_func_ratio":[0.028,0.05]},"src/punt_vox/voxd/router.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,2.05],"max_complexity":[0.0,13.0],"avg_complexity":[0.0,3.48],"module_size":[0.0,692.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
+{"ts":"2026-05-14T04:58:01Z","commit":"fac9111","files_scored":45,"files_improved":3,"files_regressed":0,"verdict":"pass","deltas":{"src/punt_vox/voxd/__init__.py":{"module_size":[136.0,122.0]},"src/punt_vox/voxd/_monolith.py":{"method_ratio":[0.435,0.587],"avg_params":[1.67,1.43],"max_complexity":[13.0,5.0],"avg_complexity":[1.98,1.33],"module_size":[1113.0,506.0],"class_to_func_ratio":[0.028,0.05]},"src/punt_vox/voxd/router.py":{"method_ratio":[0.0,1.0],"encapsulation_ratio":[0.0,1.0],"avg_params":[0.0,2.05],"max_complexity":[0.0,13.0],"avg_complexity":[0.0,3.48],"module_size":[0.0,692.0],"classes_per_module":[0.0,1.0],"class_to_func_ratio":[0.0,1.0],"init_violations":[0.0,0.0],"public_attr_violations":[0.0,0.0],"future_annotations":[0.0,1.0]}}}
