+++ .oo-baseline.json
@@ -434,7 +434,7 @@
     "avg_params": 0.0,
     "max_complexity": 0.0,
     "avg_complexity": 0.0,
-    "module_size": 136.0,
+    "module_size": 122.0,
     "classes_per_module": 0.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
@@ -442,14 +442,14 @@
     "future_annotations": 1.0
   },
   "src/punt_vox/voxd/_monolith.py": {
-    "method_ratio": 0.435,
+    "method_ratio": 0.587,
     "encapsulation_ratio": 0.375,
-    "avg_params": 1.67,
-    "max_complexity": 13.0,
-    "avg_complexity": 1.98,
-    "module_size": 1113.0,
+    "avg_params": 1.43,
+    "max_complexity": 5.0,
+    "avg_complexity": 1.33,
+    "module_size": 506.0,
     "classes_per_module": 1.0,
-    "class_to_func_ratio": 0.028,
+    "class_to_func_ratio": 0.05,
     "init_violations": 1.0,
     "public_attr_violations": 5.0,
     "future_annotations": 1.0
@@ -528,6 +528,19 @@
     "module_size": 346.0,
     "classes_per_module": 2.0,
     "class_to_func_ratio": 0.182,
+    "init_violations": 0.0,
+    "public_attr_violations": 0.0,
+    "future_annotations": 1.0
+  },
+  "src/punt_vox/voxd/router.py": {
+    "method_ratio": 1.0,
+    "encapsulation_ratio": 1.0,
+    "avg_params": 2.05,
+    "max_complexity": 13.0,
+    "avg_complexity": 3.48,
+    "module_size": 692.0,
+    "classes_per_module": 1.0,
+    "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
