+++ src/punt_vox/voxd/__init__.py
@@ -5,19 +5,12 @@
 from __future__ import annotations
 
 from punt_vox.voxd._monolith import (
-    _HANDLERS,
     DEFAULT_HOST,
     DEFAULT_PORT,
     DaemonContext,
     _apply_vibe_for_synthesis,
     _auto_track_name,
-    _handle_music_list,
-    _handle_music_next,
-    _handle_music_off,
     _handle_music_on,
-    _handle_music_play,
-    _handle_music_vibe,
-    _handle_synthesize,
     _health_payload_full,
     _health_payload_minimal,
     _health_route,
@@ -27,7 +20,6 @@ from punt_vox.voxd._monolith import (
     _music_loop,
     _synthesize_to_file,
     _try_direct_play,
-    _ws_route,
     build_app,
     cli,
     entrypoint,
@@ -67,6 +59,7 @@ from punt_vox.voxd.playback import (  # pyright: ignore[reportPrivateUsage]
     _probe_duration,
     _truncate_stderr,
 )
+from punt_vox.voxd.router import WebSocketRouter
 from punt_vox.voxd.synthesis import (  # pyright: ignore[reportPrivateUsage]
     _LOCAL_PROVIDERS,
     _PROVIDER_API_KEY_VAR,
@@ -80,7 +73,6 @@ __all__ = [
     "DEFAULT_HOST",
     "DEFAULT_PORT",
     "_DEDUP_WINDOW_SECONDS",
-    "_HANDLERS",
     "_LOCAL_PROVIDERS",
     "_MAX_STDERR_LEN",
     "_MUSIC_MAX_RETRIES",
@@ -100,18 +92,13 @@ __all__ = [
     "PlaybackQueue",
     "SynthesisPipeline",
     "TrackGenerator",
+    "WebSocketRouter",
     "_PlaybackWaitResult",
     "_apply_vibe_for_synthesis",
     "_auto_track_name",
     "_build_audio_request",
     "_config_dir",
-    "_handle_music_list",
-    "_handle_music_next",
-    "_handle_music_off",
     "_handle_music_on",
-    "_handle_music_play",
-    "_handle_music_vibe",
-    "_handle_synthesize",
     "_health_payload_full",
     "_health_payload_minimal",
     "_health_route",
@@ -129,7 +116,6 @@ __all__ = [
     "_synthesize_to_file",
     "_truncate_stderr",
     "_try_direct_play",
-    "_ws_route",
     "build_app",
     "cli",
     "entrypoint",
