+++ src/punt_vox/voxd/_monolith.py
@@ -10,28 +10,21 @@ Knows nothing about MCP, hooks, projects, sessions, or Claude Code.
 from __future__ import annotations
 
 import asyncio
-import base64
 import contextlib
-import hmac
-import json
 import logging
-import time
 from collections.abc import AsyncGenerator, Callable
 from contextlib import AbstractAsyncContextManager, asynccontextmanager
 from pathlib import Path
 from socket import socket
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, cast
 
 import typer
 import uvicorn
 from starlette.applications import Starlette
 from starlette.responses import JSONResponse
 from starlette.routing import Route, WebSocketRoute
-from starlette.websockets import WebSocket, WebSocketDisconnect, WebSocketState
 
-from punt_vox import cache as _cache_module
 from punt_vox.paths import ensure_user_dirs
-from punt_vox.providers import auto_detect_provider
 from punt_vox.voxd.chimes import ChimeResolver
 from punt_vox.voxd.config import (  # pyright: ignore[reportPrivateUsage]
     DaemonConfig,
@@ -43,23 +36,28 @@ from punt_vox.voxd.config import (  # pyright: ignore[reportPrivateUsage]
 from punt_vox.voxd.dedup import ChimeDedup, OnceDedup
 from punt_vox.voxd.health import DaemonHealth
 from punt_vox.voxd.music_scheduler import MusicScheduler
-from punt_vox.voxd.playback import PlaybackItem, PlaybackQueue
-from punt_vox.voxd.synthesis import (  # pyright: ignore[reportPrivateUsage]
-    _LOCAL_PROVIDERS,
-    SynthesisPipeline,
-)
+from punt_vox.voxd.playback import PlaybackQueue
+from punt_vox.voxd.router import WebSocketRouter
+from punt_vox.voxd.synthesis import SynthesisPipeline
 from punt_vox.voxd.track_generator import TrackGenerator
 
 if TYPE_CHECKING:
     from starlette.requests import Request
+    from starlette.websockets import WebSocket
 
 logger = logging.getLogger(__name__)
 
 DEFAULT_PORT = 8421
 DEFAULT_HOST = "127.0.0.1"
 
-# Module-level SynthesisPipeline, set by _get_pipeline() on first use.
+# Module-level SynthesisPipeline, set by main() on startup.
 _pipeline: SynthesisPipeline | None = None
+
+# Module-level TrackGenerator, set by main() at daemon startup.
+_track_generator: TrackGenerator | None = None
+
+# Module-level WebSocketRouter, set by main() at daemon startup.
+_router: WebSocketRouter | None = None
 
 
 # ---------------------------------------------------------------------------
@@ -101,33 +99,6 @@ def _get_track_generator() -> TrackGenerator:
         # Fallback for tests and handlers called before main().
         return TrackGenerator(_music_output_dir())
     return _track_generator
-
-
-# Module-level TrackGenerator, set by main() at daemon startup.
-_track_generator: TrackGenerator | None = None
-
-
-def _record_playback_result(
-    ctx: DaemonContext,
-    *,
-    path: Path,
-    rc: int,
-    elapsed: float,
-    stderr: str,
-) -> None:
-    """Update ctx.last_playback with a freshly-observed playback result.
-
-    Thin wrapper that delegates to the PlaybackQueue owned by ctx.
-    Kept as a module-level function so _try_direct_play can call it
-    without knowing about PlaybackQueue internals.
-    """
-    ctx.last_playback = {
-        "file": str(path),
-        "rc": rc,
-        "elapsed_s": round(elapsed, 4),
-        "stderr": stderr,
-        "ts": time.time(),
-    }
 
 
 # ---------------------------------------------------------------------------
@@ -179,9 +150,9 @@ class DaemonContext:
     # -- Delegation properties for PlaybackQueue -----------------------------
 
     @property
-    def playback_queue(self) -> asyncio.Queue[PlaybackItem]:
+    def playback_queue(self) -> asyncio.Queue[PlaybackQueue]:
         """Return the underlying asyncio.Queue from PlaybackQueue."""
-        return self._playback._queue
+        return self._playback._queue  # type: ignore[return-value]
 
     @property
     def last_playback(self) -> dict[str, object] | None:
@@ -285,47 +256,10 @@ class DaemonContext:
         self._music.replay = value
 
 
-_chime_resolver = ChimeResolver()
-
-
-# ---------------------------------------------------------------------------
-# Auth helper
-# ---------------------------------------------------------------------------
-
-
-def _check_auth(websocket: WebSocket, ctx: DaemonContext) -> bool:
-    """Verify the auth token from query param or first message."""
-    if ctx.auth_token is None:
-        return True  # No auth configured (tests)
-    token = websocket.query_params.get("token", "")
-    return hmac.compare_digest(token, ctx.auth_token)
-
-
-# ---------------------------------------------------------------------------
-# WebSocket message handlers
-# ---------------------------------------------------------------------------
-
-
-def _parse_optional_float(msg: dict[str, object], key: str) -> float | None:
-    """Extract an optional float field from a message dict."""
-    raw = msg.get(key)
-    if raw is None:
-        return None
-    return float(str(raw))
-
-
-def _parse_optional_int(msg: dict[str, object], key: str) -> int | None:
-    """Extract an optional int field from a message dict."""
-    raw = msg.get(key)
-    if raw is None:
-        return None
-    return int(str(raw))
-
-
-def _parse_optional_str(msg: dict[str, object], key: str) -> str | None:
-    """Extract an optional string field, returning None for empty strings."""
-    raw = str(msg.get(key, ""))
-    return raw or None
+# Backward-compatible aliases for names that moved to synthesis.py.
+# Re-exported via __init__.py and referenced by existing tests.
+_apply_vibe_for_synthesis = SynthesisPipeline.apply_vibe_for_synthesis
+_model_supports_expressive_tags = SynthesisPipeline.model_supports_expressive_tags
 
 
 def _get_pipeline(ctx: DaemonContext | None = None) -> SynthesisPipeline:
@@ -343,13 +277,27 @@ def _get_pipeline(ctx: DaemonContext | None = None) -> SynthesisPipeline:
     return _pipeline
 
 
-# Backward-compatible aliases for names that moved to synthesis.py.
-# Re-exported via __init__.py and referenced by existing tests.
-_apply_vibe_for_synthesis = SynthesisPipeline.apply_vibe_for_synthesis
-_model_supports_expressive_tags = SynthesisPipeline.model_supports_expressive_tags
+def _record_playback_result(
+    ctx: DaemonContext,
+    *,
+    path: Path,
+    rc: int,
+    elapsed: float,
+    stderr: str,
+) -> None:
+    """Update ctx.last_playback with a freshly-observed playback result."""
+    import time
+
+    ctx.last_playback = {
+        "file": str(path),
+        "rc": rc,
+        "elapsed_s": round(elapsed, 4),
+        "stderr": stderr,
+        "ts": time.time(),
+    }
 
 
-async def _synthesize_to_file(
+async def _synthesize_to_file(  # pyright: ignore[reportUnusedFunction]
     text: str,
     voice: str | None,
     provider_name: str,
@@ -384,7 +332,7 @@ async def _synthesize_to_file(
     )
 
 
-async def _try_direct_play(
+async def _try_direct_play(  # pyright: ignore[reportUnusedFunction]
     *,
     text: str,
     voice: str | None,
@@ -429,302 +377,44 @@ async def _try_direct_play(
     )
 
 
-async def _handle_synthesize(
+async def _handle_music_on(  # pyright: ignore[reportUnusedFunction]
     msg: dict[str, object],
-    websocket: WebSocket,
+    websocket: object,
     ctx: DaemonContext,
 ) -> None:
-    """Handle a 'synthesize' message: TTS + enqueue playback."""
-    request_id = str(msg.get("id", ""))
-    text = str(msg.get("text", ""))
-    if not text:
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": "empty text"}
-        )
-        return
-
-    voice = _parse_optional_str(msg, "voice")
-    provider_name = _parse_optional_str(msg, "provider") or auto_detect_provider()
-    model = _parse_optional_str(msg, "model")
-    rate = _parse_optional_int(msg, "rate")
-    language = _parse_optional_str(msg, "language")
-    vibe_tags = _parse_optional_str(msg, "vibe_tags")
-    stability = _parse_optional_float(msg, "stability")
-    similarity = _parse_optional_float(msg, "similarity")
-    style = _parse_optional_float(msg, "style")
-    speaker_boost_raw = msg.get("speaker_boost")
-    speaker_boost = bool(speaker_boost_raw) if speaker_boost_raw is not None else None
-    api_key = _parse_optional_str(msg, "api_key")
-    once = _parse_optional_int(msg, "once")
-
-    resolved_voice = voice or ""
-
-    # Opt-in dedup: only when the caller explicitly sets `once` to a
-    # positive TTL. With `once` absent, null, or 0, every request plays
-    # — the legacy always-on 5s dedup for speech was removed in vox-0e9.
-    # When we record an entry, track that fact so we can roll it back
-    # on synthesis/playback failure; otherwise a failed request would
-    # leave a zombie dedup entry that incorrectly suppresses retries.
-    dedup_recorded = False
-    if once is not None and once > 0:
-        hit = ctx.once_dedup.check_and_record(text, float(once))
-        if hit is not None:
-            logger.info(
-                "Dedup hit: id=%s text=%d chars original_played_at=%.3f "
-                "ttl_remaining=%.1fs",
-                request_id,
-                len(text),
-                hit.original_played_at,
-                hit.ttl_seconds_remaining,
-            )
-            await websocket.send_json(
-                {
-                    "type": "done",
-                    "id": request_id,
-                    "deduped": True,
-                    "original_played_at": hit.original_played_at,
-                    "ttl_seconds_remaining": hit.ttl_seconds_remaining,
-                }
-            )
-            return
-        dedup_recorded = True
-
-    def _rollback_dedup() -> None:
-        """Remove the dedup entry we recorded above, if any.
-
-        Called on every failure path between the record call and the
-        successful completion of playback. Without this, a failure
-        would leave a zombie entry in ``ctx.once_dedup._seen`` that
-        would incorrectly dedup the next retry of the same text.
-        """
-        if dedup_recorded:
-            ctx.once_dedup.rollback(text)
-
-    logger.info(
-        "Synthesize: id=%s provider=%s voice=%s chars=%d",
-        request_id,
-        provider_name,
-        resolved_voice,
-        len(text),
+    """Backward-compat wrapper: delegates to the router's handler via MusicScheduler."""
+    # Build a minimal router on the fly for backward-compat callers.
+    tg = _get_track_generator()
+    pipeline = _get_pipeline(ctx)
+    router = WebSocketRouter(
+        synthesis=pipeline,
+        playback=ctx._playback,
+        music=ctx._music,
+        chime_dedup=ctx.chime_dedup,
+        once_dedup=ctx.once_dedup,
+        chimes=ChimeResolver(),
+        health=ctx._health,
+        auth_token=ctx.auth_token,
+        track_generator=tg,
     )
-
-    # Local providers (espeak, say) play directly to the audio device,
-    # bypassing the synthesize-cache-enqueue pipeline. Cloud providers
-    # are skipped entirely so we don't pay for provider construction.
-    if provider_name in _LOCAL_PROVIDERS:
-        direct_result = await _try_direct_play(
-            text=text,
-            voice=voice,
-            provider_name=provider_name,
-            model=model,
-            language=language,
-            rate=rate,
-            vibe_tags=vibe_tags,
-            stability=stability,
-            similarity=similarity,
-            style=style,
-            speaker_boost=speaker_boost,
-            api_key=api_key,
-            ctx=ctx,
-        )
-        if isinstance(direct_result, Exception):
-            _rollback_dedup()
-            await websocket.send_json(
-                {
-                    "type": "error",
-                    "id": request_id,
-                    "message": str(direct_result),
-                }
-            )
-            return
-        if direct_result is not None:
-            if direct_result == 0:
-                await websocket.send_json({"type": "done", "id": request_id})
-            else:
-                _rollback_dedup()
-                await websocket.send_json(
-                    {
-                        "type": "error",
-                        "id": request_id,
-                        "message": f"play_directly failed with rc={direct_result}",
-                    }
-                )
-            return
-
-    try:
-        output_path = await _synthesize_to_file(
-            text,
-            voice,
-            provider_name,
-            model,
-            language,
-            rate,
-            vibe_tags,
-            stability,
-            similarity,
-            style,
-            speaker_boost=speaker_boost,
-            api_key=api_key,
-            request_id=request_id,
-        )
-    except Exception as exc:
-        _rollback_dedup()
-        logger.exception("Synthesis failed for id=%s", request_id)
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": str(exc)}
-        )
-        return
-
-    # Enqueue for playback
-    done_event = asyncio.Event()
-    await ctx.playback_queue.put(
-        PlaybackItem(path=output_path, request_id=request_id, notify=done_event)
-    )
-    await websocket.send_json({"type": "playing", "id": request_id})
-    await done_event.wait()
-    # Client may have already closed the connection after receiving 'playing'.
-    # Suppress any send failure — the audio has already played.
-    with contextlib.suppress(WebSocketDisconnect, RuntimeError):
-        await websocket.send_json({"type": "done", "id": request_id})
+    await router._handle_music_on(msg, cast("WebSocket", websocket))
 
 
-async def _handle_record(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    _ctx: DaemonContext,
-) -> None:
-    """Handle a 'record' message: TTS without playback, return audio bytes."""
-    request_id = str(msg.get("id", ""))
-    text = str(msg.get("text", ""))
-    if not text:
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": "empty text"}
-        )
-        return
-
-    voice = _parse_optional_str(msg, "voice")
-    provider_name = _parse_optional_str(msg, "provider") or auto_detect_provider()
-    model = _parse_optional_str(msg, "model")
-    rate = _parse_optional_int(msg, "rate")
-    language = _parse_optional_str(msg, "language")
-    vibe_tags = _parse_optional_str(msg, "vibe_tags")
-    stability = _parse_optional_float(msg, "stability")
-    similarity = _parse_optional_float(msg, "similarity")
-    style = _parse_optional_float(msg, "style")
-    speaker_boost_raw = msg.get("speaker_boost")
-    speaker_boost = bool(speaker_boost_raw) if speaker_boost_raw is not None else None
-    api_key = _parse_optional_str(msg, "api_key")
-
-    logger.info(
-        "Record: id=%s provider=%s voice=%s chars=%d",
-        request_id,
-        provider_name,
-        voice or "",
-        len(text),
-    )
-
-    try:
-        output_path = await _synthesize_to_file(
-            text,
-            voice,
-            provider_name,
-            model,
-            language,
-            rate,
-            vibe_tags,
-            stability,
-            similarity,
-            style,
-            speaker_boost=speaker_boost,
-            api_key=api_key,
-            request_id=request_id,
-        )
-    except Exception as exc:
-        logger.exception("Record synthesis failed for id=%s", request_id)
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": str(exc)}
-        )
-        return
-
-    audio_data = output_path.read_bytes()
-    # Only unlink tempfiles produced by a fresh synthesis. Cache-hit
-    # paths return the on-disk entry directly, and removing it would
-    # poison every subsequent identical request — including the
-    # anti-poison invariant TestCacheApiKeyBypass exercises. The
-    # CACHE_DIR lookup goes through the module (``_cache_module``)
-    # instead of a bound import so tests that monkey-patch
-    # ``punt_vox.cache.CACHE_DIR`` to a tmp dir stay in sync with the
-    # handler's view of what counts as a cache-owned path.
-    is_cache_owned = output_path.is_relative_to(_cache_module.CACHE_DIR)
-    if not is_cache_owned:
-        output_path.unlink(missing_ok=True)  # clean up temp file
-    encoded = base64.b64encode(audio_data).decode("ascii")
-    await websocket.send_json({"type": "audio", "id": request_id, "data": encoded})
+def _auto_track_name(ctx: DaemonContext) -> str:  # pyright: ignore[reportUnusedFunction]
+    """Derive a short auto-name from vibe + style + YYYYMMDD-HHMM."""
+    vibe, _ = ctx.music_vibe
+    style = ctx.music_style
+    return _get_track_generator().auto_track_name(vibe, style)
 
 
-async def _handle_chime(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    ctx: DaemonContext,
-) -> None:
-    """Handle a 'chime' message: play a bundled chime sound."""
-    signal = str(msg.get("signal", "done"))
-    path = _chime_resolver.resolve(signal)
-    if path is None:
-        logger.warning("Unknown chime signal: %s", signal)
-        await websocket.send_json(
-            {"type": "error", "id": "", "message": f"unknown chime: {signal}"}
-        )
-        return
-
-    # Chimes are always deduped with a fixed window — user explicitly
-    # confirmed this behavior is desired in vox-0e9 scoping. Unlike
-    # speech, chimes do not opt in via a `once` flag.
-    if not ctx.chime_dedup.should_play(signal):
-        logger.info("Dedup: skipping duplicate chime %s", signal)
-        await websocket.send_json({"type": "done", "id": ""})
-        return
-
-    logger.info("Chime: %s", signal)
-    done_event = asyncio.Event()
-    await ctx.playback_queue.put(
-        PlaybackItem(path=path, request_id=f"chime:{signal}", notify=done_event)
-    )
-    await websocket.send_json({"type": "playing", "id": f"chime:{signal}"})
-    await done_event.wait()
-    # Client may have already closed the connection after receiving 'playing'.
-    # Suppress any send failure — the audio has already played.
-    with contextlib.suppress(WebSocketDisconnect, RuntimeError):
-        await websocket.send_json({"type": "done", "id": f"chime:{signal}"})
+async def _kill_music_proc(ctx: DaemonContext) -> None:  # pyright: ignore[reportUnusedFunction]
+    """Kill the current music subprocess. Delegates to MusicScheduler."""
+    await ctx._music.kill_proc()
 
 
-async def _handle_voices(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    _ctx: DaemonContext,
-) -> None:
-    """Handle a 'voices' message: list available voices."""
-    from punt_vox.providers import get_provider
-
-    provider_name = _parse_optional_str(msg, "provider") or auto_detect_provider()
-
-    try:
-        provider = get_provider(provider_name, config_dir=None)
-        voice_list = await asyncio.to_thread(provider.list_voices)
-    except Exception as exc:
-        logger.exception("Voice listing failed for provider=%s", provider_name)
-        await websocket.send_json(
-            {
-                "type": "error",
-                "id": "",
-                "message": f"voice listing failed: {exc}",
-            }
-        )
-        return
-
-    await websocket.send_json(
-        {"type": "voices", "provider": provider_name, "voices": voice_list}
-    )
+async def _music_loop(ctx: DaemonContext) -> None:  # pyright: ignore[reportUnusedFunction]
+    """Background task: generate and loop music tracks. Delegates to MusicScheduler."""
+    await ctx._music.loop()
 
 
 def _health_payload_minimal(ctx: DaemonContext) -> dict[str, object]:  # pyright: ignore[reportUnusedFunction]
@@ -735,424 +425,6 @@ def _health_payload_minimal(ctx: DaemonContext) -> dict[str, object]:  # pyright
 def _health_payload_full(ctx: DaemonContext) -> dict[str, object]:  # pyright: ignore[reportUnusedFunction]
     """Return the full diagnostic health payload for authenticated callers."""
     return ctx._health.full_payload()
-
-
-async def _handle_health(
-    _msg: dict[str, object],
-    websocket: WebSocket,
-    ctx: DaemonContext,
-) -> None:
-    """Handle a 'health' message over the authenticated WebSocket."""
-    payload = ctx._health.full_payload()
-    payload["type"] = "health"
-    await websocket.send_json(payload)
-
-
-# ---------------------------------------------------------------------------
-# Music loop and handlers
-# ---------------------------------------------------------------------------
-
-
-def _slugify(text: str, max_len: int = 40) -> str:
-    """Slugify a string for use in filenames. Delegates to TrackGenerator."""
-    return TrackGenerator.slugify(text, max_len)
-
-
-async def _kill_music_proc(ctx: DaemonContext) -> None:
-    """Kill the current music subprocess. Delegates to MusicScheduler."""
-    await ctx._music.kill_proc()
-
-
-def _auto_track_name(ctx: DaemonContext) -> str:  # pyright: ignore[reportUnusedFunction]
-    """Derive a short auto-name from vibe + style + YYYYMMDD-HHMM."""
-    vibe, _ = ctx.music_vibe
-    style = ctx.music_style
-    return _get_track_generator().auto_track_name(vibe, style)
-
-
-async def _music_loop(ctx: DaemonContext) -> None:  # pyright: ignore[reportUnusedFunction]
-    """Background task: generate and loop music tracks. Delegates to MusicScheduler."""
-    await ctx._music.loop()
-
-
-async def _handle_music_on(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    ctx: DaemonContext,
-) -> None:
-    """Handle a 'music_on' message: start or transfer music ownership.
-
-    When the same owner re-sends music_on while already playing, the
-    current subprocess is kept alive and music_changed is signaled —
-    the MusicLoop's gapless handoff path generates the new track while
-    the old one keeps playing.
-
-    Ownership transfer is atomic: kill existing subprocess, update all
-    state fields, then signal MusicLoop. No interleaving.
-
-    When a ``name`` field is present and a track with that name already
-    exists on disk, the existing file is replayed without generation
-    (zero credits).
-    """
-    request_id = str(msg.get("id", ""))
-    owner_id = str(msg.get("owner_id", ""))
-    style = str(msg.get("style", ""))
-    vibe = str(msg.get("vibe", ""))
-    vibe_tags = str(msg.get("vibe_tags", ""))
-    name = str(msg.get("name", ""))
-
-    if not owner_id:
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": "owner_id is required"}
-        )
-        return
-
-    # Check for existing track by name -- skip generation if found.
-    if name:
-        safe_name = _slugify(name, max_len=60)
-        if not safe_name:
-            await websocket.send_json(
-                {"type": "error", "id": request_id, "message": "invalid track name"}
-            )
-            return
-        existing_path = _music_output_dir() / f"{safe_name}.mp3"
-        if existing_path.exists():
-            await _kill_music_proc(ctx)
-            ctx.music_mode = "on"
-            if style:
-                ctx.music_style = style
-            ctx.music_owner = owner_id
-            ctx.music_vibe = (vibe, vibe_tags)
-            ctx.music_track = existing_path
-            ctx.music_track_name = safe_name
-            ctx.music_state = "playing"
-            ctx.music_replay = True
-            ctx.music_changed.set()
-
-            logger.info(
-                "Music on (replay): owner=%s name=%s track=%s",
-                owner_id,
-                safe_name,
-                existing_path,
-            )
-            await websocket.send_json(
-                {
-                    "type": "music_on",
-                    "id": request_id,
-                    "status": "playing",
-                    "track": str(existing_path),
-                    "name": safe_name,
-                }
-            )
-            return
-
-    # When music is already playing for a different owner, kill existing
-    # playback so the new owner starts fresh.  When the *same* owner
-    # re-sends music_on (e.g. style change), skip the kill — let the
-    # MusicLoop's vibe-change path handle gapless handoff while the
-    # current track keeps looping.
-    is_already_playing = ctx.music_mode == "on" and ctx.music_proc is not None
-    if not is_already_playing or ctx.music_owner != owner_id:
-        await _kill_music_proc(ctx)
-
-    ctx.music_mode = "on"
-    if style:
-        ctx.music_style = style
-    ctx.music_owner = owner_id
-    ctx.music_vibe = (vibe, vibe_tags)
-    ctx.music_track_name = _slugify(name, max_len=60) if name else ""
-    ctx.music_replay = False
-    ctx.music_state = "generating"
-    ctx.music_changed.set()
-
-    logger.info(
-        "Music on: owner=%s style=%s vibe=%s name=%s",
-        owner_id,
-        ctx.music_style,
-        vibe,
-        ctx.music_track_name,
-    )
-    await websocket.send_json(
-        {"type": "music_on", "id": request_id, "status": "generating"}
-    )
-
-
-async def _handle_music_off(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    ctx: DaemonContext,
-) -> None:
-    """Handle a 'music_off' message: stop music playback."""
-    request_id = str(msg.get("id", ""))
-
-    await _kill_music_proc(ctx)
-    ctx.music_mode = "off"
-    ctx.music_state = "idle"
-    ctx.music_replay = False
-    ctx.music_changed.set()
-
-    logger.info("Music off")
-    await websocket.send_json(
-        {"type": "music_off", "id": request_id, "status": "stopped"}
-    )
-
-
-async def _handle_music_play(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    ctx: DaemonContext,
-) -> None:
-    """Handle a 'music_play' message: replay a saved track by name."""
-    request_id = str(msg.get("id", ""))
-    name = str(msg.get("name", ""))
-    owner_id = str(msg.get("owner_id", ""))
-
-    if not name:
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": "name is required"}
-        )
-        return
-
-    if not owner_id:
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": "owner_id is required"}
-        )
-        return
-
-    safe_name = _slugify(name, max_len=60)
-    if not safe_name:
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": "invalid track name"}
-        )
-        return
-    track_path = _music_output_dir() / f"{safe_name}.mp3"
-
-    if not track_path.exists():
-        await websocket.send_json(
-            {
-                "type": "error",
-                "id": request_id,
-                "message": f"track not found: {safe_name}",
-            }
-        )
-        return
-
-    # Kill current playback, set up replay.
-    await _kill_music_proc(ctx)
-    ctx.music_mode = "on"
-    ctx.music_owner = owner_id
-    ctx.music_track = track_path
-    ctx.music_track_name = safe_name
-    ctx.music_state = "playing"
-    ctx.music_replay = True
-    ctx.music_changed.set()
-
-    logger.info(
-        "Music play: owner=%s name=%s track=%s",
-        owner_id,
-        safe_name,
-        track_path,
-    )
-    await websocket.send_json(
-        {
-            "type": "music_play",
-            "id": request_id,
-            "status": "playing",
-            "track": str(track_path),
-            "name": safe_name,
-        }
-    )
-
-
-async def _handle_music_list(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    _ctx: DaemonContext,
-) -> None:
-    """Handle a 'music_list' message: return saved tracks with metadata."""
-    request_id = str(msg.get("id", ""))
-    tracks = _get_track_generator().list_tracks()
-
-    await websocket.send_json(
-        {
-            "type": "music_list",
-            "id": request_id,
-            "tracks": tracks,
-        }
-    )
-
-
-async def _handle_music_vibe(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    ctx: DaemonContext,
-) -> None:
-    """Handle a 'music_vibe' message: update vibe if sender is owner."""
-    request_id = str(msg.get("id", ""))
-    owner_id = str(msg.get("owner_id", ""))
-    vibe = str(msg.get("vibe", ""))
-    vibe_tags = str(msg.get("vibe_tags", ""))
-
-    if not owner_id:
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": "owner_id is required"}
-        )
-        return
-
-    if owner_id != ctx.music_owner:
-        await websocket.send_json(
-            {"type": "music_vibe", "id": request_id, "status": "ignored"}
-        )
-        return
-
-    new_vibe = (vibe, vibe_tags)
-    if new_vibe == ctx.music_vibe:
-        await websocket.send_json(
-            {"type": "music_vibe", "id": request_id, "status": "ignored"}
-        )
-        return
-
-    ctx.music_vibe = new_vibe
-    ctx.music_changed.set()
-
-    logger.info("Music vibe changed: vibe=%s tags=%s", vibe, vibe_tags)
-    await websocket.send_json(
-        {"type": "music_vibe", "id": request_id, "status": "generating"}
-    )
-
-
-async def _handle_music_next(
-    msg: dict[str, object],
-    websocket: WebSocket,
-    ctx: DaemonContext,
-) -> None:
-    """Handle a 'music_next' message: skip to a new track.
-
-    Signals MusicLoop to regenerate without changing vibe or style.
-    The current track keeps playing until the new one is ready
-    (gapless handoff via the vibe-change path).
-    """
-    request_id = str(msg.get("id", ""))
-    owner_id = str(msg.get("owner_id", ""))
-
-    if not owner_id:
-        await websocket.send_json(
-            {"type": "error", "id": request_id, "message": "owner_id is required"}
-        )
-        return
-
-    if ctx.music_mode != "on":
-        await websocket.send_json(
-            {"type": "music_next", "id": request_id, "status": "ignored"}
-        )
-        return
-
-    ctx.music_track_name = ""
-    ctx.music_replay = False
-    ctx.music_changed.set()
-
-    logger.info("Music next: owner=%s", owner_id)
-    await websocket.send_json(
-        {"type": "music_next", "id": request_id, "status": "generating"}
-    )
-
-
-# ---------------------------------------------------------------------------
-# WebSocket route
-# ---------------------------------------------------------------------------
-
-
-_HANDLERS: dict[
-    str,
-    Callable[
-        [dict[str, object], WebSocket, DaemonContext],
-        object,
-    ],
-] = {
-    "synthesize": _handle_synthesize,
-    "chime": _handle_chime,
-    "record": _handle_record,
-    "voices": _handle_voices,
-    "health": _handle_health,
-    "music_on": _handle_music_on,
-    "music_off": _handle_music_off,
-    "music_play": _handle_music_play,
-    "music_list": _handle_music_list,
-    "music_vibe": _handle_music_vibe,
-    "music_next": _handle_music_next,
-}
-
-
-async def _ws_route(websocket: WebSocket) -> None:
-    """Main WebSocket route at /ws."""
-    ctx: DaemonContext = websocket.app.state.ctx
-
-    if not _check_auth(websocket, ctx):
-        await websocket.close(code=1008)
-        return
-
-    await websocket.accept()
-    ctx.client_count += 1
-    logger.info("Client connected (total: %d)", ctx.client_count)
-
-    try:
-        while True:
-            # Preempt Starlette's RuntimeError on a peer-closed socket.
-            # After the vox-ehf fix in 4.3.0, chime/unmute clients return
-            # on the ``"playing"`` ack and close the WebSocket while this
-            # loop is still awaiting the next ``receive_text()``. The
-            # handler's trailing ``contextlib.suppress(WebSocketDisconnect,
-            # RuntimeError)`` send of the stale ``"done"`` message lands
-            # on the peer-closed socket and Starlette transitions the
-            # ``application_state`` to ``DISCONNECTED`` synchronously. If
-            # we did not check the state here, the next ``receive_text()``
-            # would raise ``RuntimeError('WebSocket is not connected. Need
-            # to call "accept" first.')`` — not ``WebSocketDisconnect`` —
-            # and fall through to ``except Exception`` below, logging a
-            # full traceback on every chime/unmute/recap. Checking the
-            # state ourselves lets us break out cleanly without widening
-            # the outer exception surface (vox-ewh).
-            if websocket.application_state != WebSocketState.CONNECTED:
-                break
-            raw = await websocket.receive_text()
-            try:
-                msg = json.loads(raw)
-            except json.JSONDecodeError:
-                await websocket.send_json(
-                    {"type": "error", "id": "", "message": "invalid JSON"}
-                )
-                continue
-
-            if not isinstance(msg, dict):
-                await websocket.send_json(
-                    {"type": "error", "id": "", "message": "expected JSON object"}
-                )
-                continue
-
-            msg_type = str(msg.get("type", ""))  # pyright: ignore[reportUnknownMemberType,reportUnknownArgumentType]
-            handler = _HANDLERS.get(msg_type)
-            if handler is None:
-                msg_id = str(msg.get("id", ""))  # pyright: ignore[reportUnknownMemberType,reportUnknownArgumentType]
-                await websocket.send_json(
-                    {
-                        "type": "error",
-                        "id": msg_id,
-                        "message": f"unknown message type: {msg_type}",
-                    }
-                )
-                continue
-
-            # Each message handler is awaited in the receive loop.
-            # Multiple clients are concurrent (each has its own receive loop),
-            # but messages from a single client are processed sequentially.
-            await handler(msg, websocket, ctx)  # type: ignore[misc]
-    except WebSocketDisconnect:
-        pass
-    except Exception:
-        logger.exception("WebSocket error")
-    finally:
-        ctx.client_count -= 1
-        logger.info("Client disconnected (total: %d)", ctx.client_count)
 
 
 # ---------------------------------------------------------------------------
@@ -1175,6 +447,7 @@ def build_app(
     ctx: DaemonContext | None = None,
     *,
     lifespan: (Callable[[Starlette], AbstractAsyncContextManager[None]] | None) = None,
+    router: WebSocketRouter | None = None,
 ) -> Starlette:
     """Build the Starlette ASGI application.
 
@@ -1184,9 +457,25 @@ def build_app(
     if ctx is None:
         ctx = DaemonContext()
 
+    if router is None:
+        # Create a minimal router for backward compatibility (tests).
+        pipeline = _pipeline or SynthesisPipeline(playback_mutex=ctx._playback.mutex)
+        tg = _track_generator or TrackGenerator(_music_output_dir())
+        router = WebSocketRouter(
+            synthesis=pipeline,
+            playback=ctx._playback,
+            music=ctx._music,
+            chime_dedup=ctx.chime_dedup,
+            once_dedup=ctx.once_dedup,
+            chimes=ChimeResolver(),
+            health=ctx._health,
+            auth_token=ctx.auth_token,
+            track_generator=tg,
+        )
+
     routes: list[Route | WebSocketRoute] = [
         Route("/health", _health_route, methods=["GET"]),
-        WebSocketRoute("/ws", _ws_route),
+        WebSocketRoute("/ws", router.handle_connection),
     ]
 
     app = Starlette(routes=routes, lifespan=lifespan)
@@ -1209,11 +498,6 @@ def main(
     ),
 ) -> None:
     """Start the voxd audio server daemon."""
-    # Create (or tighten) per-user state dirs before anything else
-    # touches the filesystem. ``ensure_user_dirs`` forces mode 0700 on
-    # ``~/.punt-labs/vox`` and its ``logs``/``run``/``cache``
-    # subdirectories, including pre-existing dirs that were created
-    # under a looser umask in earlier versions.
     ensure_user_dirs()
 
     daemon_cfg = DaemonConfig(
@@ -1245,7 +529,7 @@ def main(
     # Music scheduler owns the background loop and all music state.
     scheduler = MusicScheduler(_track_generator)
     playback = PlaybackQueue()
-    health = DaemonHealth(playback, lambda: ctx.client_count, port)
+    health = DaemonHealth(playback, lambda: router.client_count, port)
     ctx = DaemonContext(
         auth_token=auth_token,
         port=port,
@@ -1257,6 +541,21 @@ def main(
     # Initialize synthesis pipeline eagerly with the real playback mutex.
     global _pipeline
     _pipeline = SynthesisPipeline(playback_mutex=ctx._playback.mutex)
+
+    # Build the WebSocket router with all dependencies.
+    global _router
+    router = WebSocketRouter(
+        synthesis=_pipeline,
+        playback=playback,
+        music=scheduler,
+        chime_dedup=ctx.chime_dedup,
+        once_dedup=ctx.once_dedup,
+        chimes=ChimeResolver(),
+        health=health,
+        auth_token=auth_token,
+        track_generator=_track_generator,
+    )
+    _router = router
 
     logger.info("Starting voxd on %s:%d", host, port)
 
@@ -1282,7 +581,7 @@ def main(
             daemon_cfg.remove_port_file()
             logger.info("voxd stopped")
 
-    app = build_app(ctx, lifespan=lifespan)
+    app = build_app(ctx, lifespan=lifespan, router=router)
 
     if host not in ("127.0.0.1", "::1", "localhost"):
         logger.warning(
